+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/command/EntityPoolIterationHandler.java
@@ -2,10 +2,10 @@ package org.seasar.doma.jdbc.criteria.command;
 
 import java.util.List;
 import java.util.Objects;
-import org.seasar.doma.def.EntityDef;
 import org.seasar.doma.internal.jdbc.command.AbstractIterationHandler;
 import org.seasar.doma.internal.jdbc.command.ResultListCallback;
 import org.seasar.doma.jdbc.ObjectProvider;
+import org.seasar.doma.jdbc.criteria.def.EntityDef;
 import org.seasar.doma.jdbc.query.SelectQuery;
 
 public class EntityPoolIterationHandler
