+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/command/EntityPoolProvider.java
@@ -13,12 +13,12 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import org.seasar.doma.DomaNullPointerException;
-import org.seasar.doma.def.EntityDef;
 import org.seasar.doma.internal.jdbc.command.JdbcValueGetter;
 import org.seasar.doma.jdbc.JdbcMappable;
 import org.seasar.doma.jdbc.JdbcMappingFunction;
 import org.seasar.doma.jdbc.JdbcMappingVisitor;
 import org.seasar.doma.jdbc.ObjectProvider;
+import org.seasar.doma.jdbc.criteria.def.EntityDef;
 import org.seasar.doma.jdbc.entity.EntityPropertyType;
 import org.seasar.doma.jdbc.entity.EntityType;
 import org.seasar.doma.jdbc.entity.Property;
