+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/command/MappedObjectIterationHandler.java
@@ -3,10 +3,10 @@ package org.seasar.doma.jdbc.criteria.command;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Function;
-import org.seasar.doma.def.PropertyDef;
 import org.seasar.doma.internal.jdbc.command.AbstractIterationHandler;
 import org.seasar.doma.internal.jdbc.command.ResultListCallback;
 import org.seasar.doma.jdbc.ObjectProvider;
+import org.seasar.doma.jdbc.criteria.def.PropertyDef;
 import org.seasar.doma.jdbc.criteria.statement.Row;
 import org.seasar.doma.jdbc.query.SelectQuery;
