+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/command/MappedObjectProvider.java
@@ -7,9 +7,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Function;
-import org.seasar.doma.def.PropertyDef;
 import org.seasar.doma.internal.jdbc.command.AbstractObjectProvider;
 import org.seasar.doma.jdbc.JdbcMappingVisitor;
+import org.seasar.doma.jdbc.criteria.def.PropertyDef;
 import org.seasar.doma.jdbc.criteria.statement.Row;
 import org.seasar.doma.jdbc.entity.Property;
 import org.seasar.doma.jdbc.query.Query;
