+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/declaration/ValuesDeclaration.java
@@ -1,8 +1,8 @@
 package org.seasar.doma.jdbc.criteria.declaration;
 
 import java.util.Objects;
-import org.seasar.doma.def.PropertyDef;
 import org.seasar.doma.jdbc.criteria.context.InsertContext;
+import org.seasar.doma.jdbc.criteria.def.PropertyDef;
 
 public class ValuesDeclaration {
