+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/def/DefaultPropertyDef.java
@@ -1,4 +1,4 @@
-package org.seasar.doma.def;
+package org.seasar.doma.jdbc.criteria.def;
 
 import org.seasar.doma.jdbc.entity.EntityPropertyType;
 import org.seasar.doma.jdbc.entity.EntityType;
