+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/def/EmbeddableDef.java
@@ -1,4 +1,4 @@
-package org.seasar.doma.def;
+package org.seasar.doma.jdbc.criteria.def;
 
 import java.util.List;
