+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/query/InsertBuilder.java
@@ -2,7 +2,6 @@ package org.seasar.doma.jdbc.criteria.query;
 
 import java.util.Objects;
 import java.util.function.Function;
-import org.seasar.doma.def.EntityDef;
 import org.seasar.doma.internal.jdbc.sql.PreparedSqlBuilder;
 import org.seasar.doma.jdbc.Config;
 import org.seasar.doma.jdbc.PreparedSql;
@@ -10,6 +9,7 @@ import org.seasar.doma.jdbc.SqlKind;
 import org.seasar.doma.jdbc.SqlLogType;
 import org.seasar.doma.jdbc.criteria.context.InsertContext;
 import org.seasar.doma.jdbc.criteria.context.Operand;
+import org.seasar.doma.jdbc.criteria.def.EntityDef;
 
 public class InsertBuilder {
   private final Config config;
