+++ doma-core/src/main/java/org/seasar/doma/jdbc/criteria/query/UpdateBuilder.java
@@ -4,7 +4,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.function.Function;
-import org.seasar.doma.def.EntityDef;
 import org.seasar.doma.internal.jdbc.sql.PreparedSqlBuilder;
 import org.seasar.doma.jdbc.Config;
 import org.seasar.doma.jdbc.PreparedSql;
@@ -13,6 +12,7 @@ import org.seasar.doma.jdbc.SqlLogType;
 import org.seasar.doma.jdbc.criteria.context.Criterion;
 import org.seasar.doma.jdbc.criteria.context.Operand;
 import org.seasar.doma.jdbc.criteria.context.UpdateContext;
+import org.seasar.doma.jdbc.criteria.def.EntityDef;
 
 public class UpdateBuilder {
   private final UpdateContext context;
