+++ doma-core/src/test/java/org/seasar/doma/jdbc/criteria/entity/NoIdEmp_.java
@@ -3,9 +3,9 @@ package org.seasar.doma.jdbc.criteria.entity;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
-import org.seasar.doma.def.DefaultPropertyDef;
-import org.seasar.doma.def.EntityDef;
-import org.seasar.doma.def.PropertyDef;
+import org.seasar.doma.jdbc.criteria.def.DefaultPropertyDef;
+import org.seasar.doma.jdbc.criteria.def.EntityDef;
+import org.seasar.doma.jdbc.criteria.def.PropertyDef;
 
 public class NoIdEmp_ implements EntityDef<NoIdEmp> {
