+++ doma-processor/src/test/java/org/seasar/doma/internal/apt/processor/entitydesc/EmpInfo_.java
@@ -3,9 +3,9 @@ package org.seasar.doma.internal.apt.processor.entitydesc;
 import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.List;
-import org.seasar.doma.def.DefaultPropertyDef;
-import org.seasar.doma.def.EmbeddableDef;
-import org.seasar.doma.def.PropertyDef;
+import org.seasar.doma.jdbc.criteria.def.DefaultPropertyDef;
+import org.seasar.doma.jdbc.criteria.def.EmbeddableDef;
+import org.seasar.doma.jdbc.criteria.def.PropertyDef;
 import org.seasar.doma.jdbc.entity.EntityType;
 
 public class EmpInfo_ implements EmbeddableDef {
